requirements.txt
fastapi
uvicorn
psycopg2
sqlalchemy
sqlmodel
email-validator
passlib
bcrypt=3.2.2
pyjwt
python-multipart
alembic
pydantic-settings
